Added support PHP 7.0:
- renamed `Validator\Rule\Float` to `FloatRule`
- renamed `Validator\Rule\String` to `StringRule`
Fixed codestyle

+++ src/Session/Adapter/AbstractAdapter.php
@@ -51,7 +51,7 @@ abstract class AbstractAdapter
     public function open($savePath, $sessionName)
     {
         $this->prefix = $sessionName . ':';
-        $this->ttl = ini_get('session.gc_maxlifetime');
+        $this->ttl = (int) ini_get('session.gc_maxlifetime');
 
         // No more action necessary because connection is injected
         // in constructor and arguments are not applicable.
